+++ src/routes/(v2)/v2/Navigation.svelte
@@ -80,7 +80,8 @@
 		'timeproxy',
 		'tainted-null-random',
 		'issue-588',
-		'customrequest-validate'
+		'customrequest-validate',
+		'issue-655'
 	].sort();
 </script>