+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Migrated Valibot adapter to use the official `@valibot/to-json-schema` package. [#668](https://github.com/ciscoheat/sveltekit-superforms/pull/668)
+
 ## [2.28.1] - 2025-10-19
 
 ### Fixed